Move Sampler abstract class to the main package.
