+++ api/src/main/java/io/opencensus/trace/Sampler.java
@@ -11,11 +11,13 @@
  * limitations under the License.
  */
 
-package io.opencensus.trace.base;
+package io.opencensus.trace;
 
 import io.opencensus.trace.Span;
 import io.opencensus.trace.SpanContext;
 
+import io.opencensus.trace.base.SpanId;
+import io.opencensus.trace.base.TraceId;
 import java.util.List;
 import javax.annotation.Nullable;
