+++ api/src/main/java/io/opencensus/trace/SpanBuilder.java
@@ -17,7 +17,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import io.opencensus.common.NonThrowingCloseable;
 import io.opencensus.trace.base.EndSpanOptions;
-import io.opencensus.trace.base.Sampler;
 import java.util.List;
 import javax.annotation.Nullable;
