+++ api/src/main/java/io/opencensus/trace/samplers/AlwaysSampleSampler.java
@@ -13,23 +13,20 @@
 
 package io.opencensus.trace.samplers;
 
+import io.opencensus.trace.Sampler;
 import io.opencensus.trace.Span;
 import io.opencensus.trace.SpanContext;
-import io.opencensus.trace.base.Sampler;
 import io.opencensus.trace.base.SpanId;
 import io.opencensus.trace.base.TraceId;
 import java.util.List;
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.Immutable;
 
-/**
- * Sampler that always makes a "yes" decision on {@link Span} sampling.
- */
+/** Sampler that always makes a "yes" decision on {@link Span} sampling. */
 @Immutable
 final class AlwaysSampleSampler extends Sampler {
 
-  AlwaysSampleSampler() {
-  }
+  AlwaysSampleSampler() {}
 
   // Returns always makes a "yes" decision on {@link Span} sampling.
   @Override
